+++ src/CrossValidation/Metrics/Metric.php
@@ -2,6 +2,7 @@
 
 namespace Rubix\ML\CrossValidation\Metrics;
 
+use Rubix\ML\Tuple;
 use Stringable;
 
 interface Metric extends Stringable
@@ -9,9 +10,9 @@ interface Metric extends Stringable
     /**
      * Return a tuple of the min and max output value for this metric.
      *
-     * @return array{float,float}
+     * @return \Rubix\ML\Tuple{float,float}
      */
-    public function range() : array;
+    public function range() : Tuple;
 
     /**
      * The estimator types that this metric is compatible with.