+++ tests/CrossValidation/MonteCarloTest.php
@@ -74,7 +74,7 @@ class MonteCarloTest extends TestCase
      */
     public function test() : void
     {
-        [$min, $max] = $this->metric->range();
+        [$min, $max] = $this->metric->range()->list();
 
         $dataset = $this->generator->generate(self::DATASET_SIZE);