+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - Added Dataset general sort() using callback
     - Confusion Matrix classes no longer selectable
     - Remove Recursive Feature Eliminator transformer
+    - Metric range() now returns a Tuple object
 
 - 1.0.0-beta1
     - Added variance smoothing to Gaussian NB, Mixture, and MLE